# src/modules/notes/note.entity.ts
import { Entity, Column, PrimaryGeneratedColumn, ManyToOne, JoinTable } from 'typeorm';
import {User} from '../user/user.entity';

@Entity()
export class Note{
  @PrimaryGeneratedColumn("uuid")
  id: string

  @Column()
  content: string

  @ManyToOne((type => User),(user)=>user.notes,{
    orphanedRowAction: 'delete',
  })
  user: User

  @ManyToOne((type) => User,(user)=>user.myNotes,{
    orphanedRowAction: 'delete',
  })
  author: User;
}